requirements.txt
# Web Framework
fastapi
uvicorn
jinja2

# HTTP & Scraping
requests
beautifulsoup4

# Database
sqlalchemy
psycopg2-binary

# LLM
google-generativeai
python-dotenv

# Scheduling
apscheduler
